+++ packages/server/lib/crons/delete/deleteSyncConfigData.ts
