+++ packages/server/lib/crons/delete/deleteSyncData.ts
